refactor: lobby,profileをTabBarWidgetに移植
