--- lib/pages/lobby_page.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:supabase_flutter/supabase_flutter.dart';
-
-class LobbyPage extends StatelessWidget {
-  const LobbyPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final supabase = Supabase.instance.client;
-    const headingTextStyle = TextStyle(fontSize: 36, color: Colors.black);
-
-    return Scaffold(
-      appBar: AppBar(
-        actions: [
-          IconButton(
-            iconSize: 50,
-            icon: Icon(Icons.person),
-            onPressed: () {
-              Navigator.pushNamed(context, '/profile');
-            },
-          ),
-          IconButton(
-            iconSize: 50,
-            icon: Icon(Icons.logout),
-            onPressed: () {
-              supabase.auth.signOut();
-              Navigator.pushNamed(context, '/signin');
-            },
-          ),
-        ],
-      ),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {},
-          child: const Column(
-            mainAxisSize: MainAxisSize.min,
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Text('ラジオ体操に', style: headingTextStyle),
-              Text("参加する", style: headingTextStyle),
-              SizedBox(height: 20),
-              Text('212人が参加中...'),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
