+++ lib/pages/screens/home_screen.dart
@@ -1,16 +1,42 @@
 import 'package:flutter/material.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final supabase = Supabase.instance.client;
+    const headingTextStyle = TextStyle(fontSize: 36, color: Colors.black);
+
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Home'),
+        actions: [
+          IconButton(
+            iconSize: 50,
+            icon: Icon(Icons.logout),
+            onPressed: () {
+              supabase.auth.signOut();
+              Navigator.pushNamed(context, '/signin');
+            },
+          ),
+        ],
       ),
-      body:
-          const Center(child: Text('Home画面', style: TextStyle(fontSize: 32.0))),
+      body: Center(
+        child: ElevatedButton(
+          onPressed: () {},
+          child: const Column(
+            mainAxisSize: MainAxisSize.min,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text('ラジオ体操に', style: headingTextStyle),
+              Text("参加する", style: headingTextStyle),
+              SizedBox(height: 20),
+              Text('212人が参加中...'),
+            ],
+          ),
+        ),
+      ),
     );
   }
 }
